move migration_targets function
